+++ demo/routes/Breadcrumb/index.js
@@ -1,0 +1,8 @@
+export default {
+  path: 'breadcrumb',
+  getComponent(nextState, cb) {
+    require.ensure([], (require) => {
+      cb(null, require('./components/Breadcrumb').default);
+    });
+  }
+};