+++ demo/routes/Clipboard/index.js
@@ -1,0 +1,8 @@
+export default {
+  path: 'clipboard',
+  getComponent(nextState, cb) {
+    require.ensure([], (require) => {
+      cb(null, require('./components/Clipboard').default);
+    });
+  }
+};